requirements.txt
absl-py==2.0.0
asgiref==3.7.2
astunparse==1.6.3
azure-storage-blob==12.18.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.1
Django==4.2.6
django-cors-headers==4.3.0
flatbuffers==23.5.26
fonttools==4.43.1
gast==0.5.4
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.9.0
idna==3.4
isodate==0.6.1
joblib==1.3.2
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
ml-dtypes==0.2.0
numpy
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas
Pillow
protobuf==4.24.4
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.1
scipy
six==1.16.0
sqlparse==0.4.4
tensorboard
tensorboard-data-server
tensorflow-cpu
tensorflow-estimator
tensorflow-intel
tensorflow-io-gcs-filesystem
termcolor
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
Werkzeug
whitenoise
wrapt==1.14.1
python-decouple